+++ test/instructions/from.test.ts
@@ -13,6 +13,7 @@ describe("FROM", () => {
         let from = dockerfile.getFROMs()[0];
         assert.equal(from.getImage(), null);
         assert.equal(from.getImageName(), null);
+        assert.equal(from.getImageNameRange(), null);
         assert.equal(from.getImageTagRange(), null);
         assert.equal(from.getImageDigestRange(), null);
         assert.equal(from.getBuildStage(), null);
@@ -23,6 +24,7 @@ describe("FROM", () => {
         let from = dockerfile.getFROMs()[0];
         assert.equal(from.getImage(), "node");
         assert.equal(from.getImageName(), "node");
+        assertRange(from.getImageNameRange(), 0, 5, 0, 9);
         assert.equal(from.getImageTagRange(), null);
         assert.equal(from.getImageDigestRange(), null);
         assert.equal(from.getBuildStage(), null);
@@ -33,6 +35,7 @@ describe("FROM", () => {
         let from = dockerfile.getFROMs()[0];
         assert.equal(from.getImage(), "node:");
         assert.equal(from.getImageName(), "node");
+        assertRange(from.getImageNameRange(), 0, 5, 0, 9);
         assertRange(from.getImageTagRange(), 0, 10, 0, 10);
         assert.equal(from.getImageDigestRange(), null);
         assert.equal(from.getBuildStage(), null);
@@ -43,6 +46,7 @@ describe("FROM", () => {
         let from = dockerfile.getFROMs()[0];
         assert.equal(from.getImage(), "node:alpine");
         assert.equal(from.getImageName(), "node");
+        assertRange(from.getImageNameRange(), 0, 5, 0, 9);
         assertRange(from.getImageTagRange(), 0, 10, 0, 16);
         assert.equal(from.getImageDigestRange(), null);
         assert.equal(from.getBuildStage(), null);
@@ -53,6 +57,7 @@ describe("FROM", () => {
         let from = dockerfile.getFROMs()[0];
         assert.equal(from.getImage(), "node@");
         assert.equal(from.getImageName(), "node");
+        assertRange(from.getImageNameRange(), 0, 5, 0, 9);
         assert.equal(from.getImageTagRange(), null);
         assertRange(from.getImageDigestRange(), 0, 10, 0, 10);
         assert.equal(from.getBuildStage(), null);
@@ -63,6 +68,7 @@ describe("FROM", () => {
         let from = dockerfile.getFROMs()[0];
         assert.equal(from.getImage(), "node@sha256:613685c22f65d01f2264bdd49b8a336488e14faf29f3ff9b6bf76a4da23c4700");
         assert.equal(from.getImageName(), "node");
+        assertRange(from.getImageNameRange(), 0, 5, 0, 9);
         assert.equal(from.getImageTagRange(), null);
         assertRange(from.getImageDigestRange(), 0, 10, 0, 81);
         assert.equal(from.getBuildStage(), null);
@@ -73,6 +79,7 @@ describe("FROM", () => {
         let from = dockerfile.getFROMs()[0];
         assert.equal(from.getImage(), "node");
         assert.equal(from.getImageName(), "node");
+        assertRange(from.getImageNameRange(), 0, 5, 0, 9);
         assert.equal(from.getImageTagRange(), null);
         assert.equal(from.getImageDigestRange(), null);
         assert.equal(from.getBuildStage(), null);
@@ -83,6 +90,7 @@ describe("FROM", () => {
         let from = dockerfile.getFROMs()[0];
         assert.equal(from.getImage(), "node");
         assert.equal(from.getImageName(), "node");
+        assertRange(from.getImageNameRange(), 0, 5, 0, 9);
         assert.equal(from.getImageTagRange(), null);
         assert.equal(from.getImageDigestRange(), null);
         assert.equal(from.getBuildStage(), "stage");